requirements.pip.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --all-extras --no-emit-index-url --output-file=requirements.pip.txt pyproject.toml
#
alabaster==0.7.13
    # via sphinx
attrs==23.1.0
    # via jsonschema
babel==2.12.1
    # via sphinx
beautifulsoup4==4.12.2
    # via furo
build==0.10.0
    # via
    #   hat-doit
    #   pip-tools
certifi==2023.7.22
    # via requests
charset-normalizer==3.2.0
    # via requests
click==8.1.6
    # via pip-tools
cloudpickle==2.2.1
    # via doit
coverage[toml]==7.2.7
    # via pytest-cov
docutils==0.20.1
    # via sphinx
doit==0.36.0
    # via hat-doit
flake8==6.1.0
    # via hat-doit
furo==2023.7.26
    # via hat-doit
hat-doit==0.14.4
    # via dtviz (pyproject.toml)
hat-json==0.5.19
    # via dtviz (pyproject.toml)
hat-util==0.6.10
    # via hat-json
idna==3.4
    # via requests
imagesize==1.4.1
    # via sphinx
importlib-metadata==6.8.0
    # via doit
iniconfig==2.0.0
    # via pytest
jinja2==3.1.2
    # via
    #   pdoc
    #   sphinx
jsonpatch==1.33
    # via hat-json
jsonpointer==2.4
    # via jsonpatch
jsonschema==4.17.3
    # via hat-json
markupsafe==2.1.3
    # via
    #   jinja2
    #   pdoc
mccabe==0.7.0
    # via flake8
packaging==23.1
    # via
    #   build
    #   dtviz (pyproject.toml)
    #   hat-doit
    #   pytest
    #   sphinx
pdoc==14.0.0
    # via hat-doit
pip-tools==7.2.0
    # via hat-doit
pluggy==1.2.0
    # via pytest
pycodestyle==2.11.0
    # via flake8
pyflakes==3.1.0
    # via flake8
pygments==2.16.1
    # via
    #   furo
    #   pdoc
    #   sphinx
pyproject-hooks==1.0.0
    # via build
pyrsistent==0.19.3
    # via jsonschema
pytest==7.4.0
    # via
    #   hat-doit
    #   pytest-asyncio
    #   pytest-cov
    #   pytest-timeout
pytest-asyncio==0.21.1
    # via hat-doit
pytest-cov==4.1.0
    # via hat-doit
pytest-timeout==2.1.0
    # via hat-doit
pyyaml==6.0.1
    # via hat-json
requests==2.31.0
    # via sphinx
snowballstemmer==2.2.0
    # via sphinx
soupsieve==2.4.1
    # via beautifulsoup4
sphinx==7.1.2
    # via
    #   furo
    #   hat-doit
    #   sphinx-basic-ng
sphinx-basic-ng==1.0.0b2
    # via furo
sphinxcontrib-applehelp==1.0.4
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.1
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
tomli==2.0.1
    # via
    #   hat-doit
    #   hat-json
tomli-w==1.0.0
    # via
    #   hat-doit
    #   hat-json
urllib3==2.0.4
    # via requests
wheel==0.41.1
    # via
    #   hat-doit
    #   pip-tools
zipp==3.16.2
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
